Create timeAgo util
