+++ src/utils/time-ago.ts
@@ -1,0 +1,19 @@
+const timeAgo = (timestamp: number): string => {
+  const now = Date.now();
+  const diff = now - timestamp;
+  const seconds = Math.floor(diff / 1000);
+  if (seconds < 60) {
+    return `${seconds}s`;
+  } else if (seconds < 3600) {
+    const minutes = Math.floor(seconds / 60);
+    return `${minutes}m`;
+  } else if (seconds < 86400) {
+    const hours = Math.floor(seconds / 3600);
+    return `${hours}h`;
+  } else {
+    const days = Math.floor(seconds / 86400);
+    return `${days}d`;
+  }
+};
+
+export default timeAgo;
